优化

+++ light-game/src/main/java/com/xhn/light/game/entity/ScoreEntity.java
@@ -28,7 +28,7 @@ public class ScoreEntity implements Serializable {
 	/**
 	 * 游戏分数
 	 */
-	private Integer gameScore;
+	private Double gameScore;
 	/**
 	 * 评分类型
 	 */
